fix: 나중에 보내는 선물박스의 gift null check 추가

+++ packy-api/src/main/java/com/dilly/gift/application/GiftBoxService.java
@@ -302,7 +302,8 @@ public class GiftBoxService {
                     fileService.deleteFile(photo.getImgUrl());
                     photoWriter.delete(photo);
                 });
-                if (giftBox.getGift().getGiftType().equals(GiftType.PHOTO)) {
+                if (giftBox.getGift() != null && giftBox.getGift().getGiftType()
+                    .equals(GiftType.PHOTO)) {
                     fileService.deleteFile(giftBox.getGift().getGiftUrl());
                 }
                 giftBox.getGiftBoxStickers().forEach(giftBoxStickerWriter::delete);
